refactor(blog): isolate strict sanitizer for Table, revert shared sanitizeHtml
Revert blog/utils/sanitizeHtml.ts to its original form so the many blocks that
import it are unaffected. Move the hardened denylist logic into a dedicated
blog/utils/sanitizeHtmlStrict.ts and use it only in Table.tsx.

sanitizeHtmlStrict removes dangerous elements with their content (iframe,
object, embed, svg, form, …), strips on*/srcdoc/style attributes, and
neutralizes javascript:/data:/vbscript: protocols in url-bearing attributes —
handling quoted and unquoted values identically and decoding/normalizing the
value first to catch obfuscated schemes.
